+++ orchestration/hca_orchestration/repositories/dev_repository.py
@@ -10,12 +10,11 @@ from dagster_utils.resources.data_repo.jade_data_repo import jade_data_repo_clie
 from dagster_utils.resources.google_storage import google_storage_client
 from dagster_utils.resources.slack import live_slack_client
 
-from hca_orchestration.contrib.dagster import configure_partitions_for_pipeline
 from hca_orchestration.config import preconfigure_resource_for_mode
 from hca_orchestration.config.dcp_release.dcp_release import run_config_for_dcp_release_partition
 from hca_orchestration.config.dev_refresh.dev_refresh import run_config_for_per_project_dataset_partition, \
     run_config_for_cut_snapshot_partition
-from hca_orchestration.pipelines import copy_project
+from hca_orchestration.contrib.dagster import configure_partitions_for_pipeline
 from hca_orchestration.pipelines.cut_snapshot import legacy_cut_snapshot_job, cut_project_snapshot_job
 from hca_orchestration.pipelines.load_hca import load_hca
 from hca_orchestration.pipelines.validate_ingress import validate_ingress_graph, staging_area_validator, \
@@ -24,9 +23,9 @@ from hca_orchestration.resources import bigquery_service
 from hca_orchestration.resources import load_tag
 from hca_orchestration.resources.config.dagit import dagit_config
 from hca_orchestration.resources.config.scratch import scratch_config
-from hca_orchestration.resources.config.target_hca_dataset import target_hca_dataset, build_new_target_hca_dataset
+from hca_orchestration.resources.config.target_hca_dataset import target_hca_dataset
 from hca_orchestration.resources.data_repo_service import data_repo_service
-from hca_orchestration.resources.hca_project_config import hca_project_copying_config
+from hca_orchestration.repositories.common import copy_project_to_new_dataset_job
 
 
 def validate_ingress_job() -> PipelineDefinition:
@@ -56,22 +55,6 @@ def load_hca_job() -> PipelineDefinition:
             "dagit_config": preconfigure_resource_for_mode(dagit_config, "dev")
         }
     )
-
-
-def copy_project_to_new_dataset_job() -> PipelineDefinition:
-    return copy_project.to_job(
-        name="copy_project_to_new_dataset",
-        resource_defs={
-            "bigquery_client": bigquery_client,
-            "data_repo_client": jade_data_repo_client,
-            "gcs": google_storage_client,
-            "scratch_config": scratch_config,
-            "bigquery_service": bigquery_service,
-            "hca_project_copying_config": hca_project_copying_config,
-            "target_hca_dataset": build_new_target_hca_dataset,
-            "load_tag": load_tag,
-            "data_repo_service": data_repo_service,
-        })
 
 
 @repository
