+++ orchestration/hca_orchestration/repositories/local_repository.py
@@ -8,13 +8,15 @@ from dagster_utils.resources.beam.local_beam_runner import local_beam_runner
 from dagster_utils.resources.bigquery import bigquery_client
 from dagster_utils.resources.data_repo.jade_data_repo import jade_data_repo_client
 from dagster_utils.resources.google_storage import google_storage_client
-from dagster_utils.resources.sam import sam_client
-from dagster_utils.resources.slack import console_slack_client, live_slack_client
+from dagster_utils.resources.slack import console_slack_client
 
-from hca_orchestration.contrib.dagster import configure_partitions_for_pipeline
+from hca_orchestration.config.dcp_release.dcp_release import run_config_for_dcp_release_partition
+from hca_orchestration.config.dev_refresh.dev_refresh import run_config_for_per_project_dataset_partition, \
+    run_config_for_cut_snapshot_partition
 from hca_orchestration.config import preconfigure_resource_for_mode
+from hca_orchestration.contrib.dagster import configure_partitions_for_pipeline
+from hca_orchestration.pipelines.cut_snapshot import cut_project_snapshot_job, legacy_cut_snapshot_job
 from hca_orchestration.pipelines.load_hca import load_hca
-from hca_orchestration.pipelines.cut_snapshot import cut_snapshot
 from hca_orchestration.pipelines.validate_ingress import run_config_for_validation_ingress_partition, \
     validate_ingress_graph, staging_area_validator
 from hca_orchestration.resources import load_tag, bigquery_service
@@ -22,7 +24,7 @@ from hca_orchestration.resources.config.dagit import dagit_config
 from hca_orchestration.resources.config.scratch import scratch_config
 from hca_orchestration.resources.config.target_hca_dataset import target_hca_dataset
 from hca_orchestration.resources.data_repo_service import data_repo_service
-from hca_orchestration.resources.config.data_repo import snapshot_creation_config, hca_manage_config
+from hca_orchestration.repositories.common import copy_project_to_new_dataset_job
 
 
 def validate_ingress_job() -> PipelineDefinition:
@@ -55,28 +57,19 @@ def load_hca_job() -> PipelineDefinition:
     )
 
 
-def cut_snapshot_job() -> PipelineDefinition:
-    return cut_snapshot.to_job(
-        resource_defs={
-            "data_repo_client": preconfigure_resource_for_mode(jade_data_repo_client, "dev"),
-            "data_repo_service": data_repo_service,
-            "gcs": google_storage_client,
-            "hca_manage_config": preconfigure_resource_for_mode(hca_manage_config, "dev"),
-            "sam_client": preconfigure_resource_for_mode(sam_client, "dev"),
-            "slack": preconfigure_resource_for_mode(live_slack_client, "local"),
-            "snapshot_config": snapshot_creation_config,
-            "dagit_config": preconfigure_resource_for_mode(dagit_config, "local"),
-        },
-        executor_def=in_process_executor
-    )
-
-
 @repository
 def all_jobs() -> list[PipelineDefinition]:
     jobs = [
+        copy_project_to_new_dataset_job(),
+        cut_project_snapshot_job("dev", "dev", "monster-dev@dev.test.firecloud.org"),
+        legacy_cut_snapshot_job("dev", "monster-dev@dev.test.firecloud.org"),
         load_hca_job(),
-        cut_snapshot_job(),
         validate_ingress_job()
     ]
+    jobs += configure_partitions_for_pipeline("copy_project_to_new_dataset",
+                                              run_config_for_per_project_dataset_partition)
+    jobs += configure_partitions_for_pipeline("cut_snapshot", run_config_for_cut_snapshot_partition)
+    jobs += configure_partitions_for_pipeline("load_hca", run_config_for_dcp_release_partition)
     jobs += configure_partitions_for_pipeline("validate_ingress", run_config_for_validation_ingress_partition)
+
     return jobs
