+++ orchestration/hca_orchestration/repositories/prod_repository.py
@@ -10,6 +10,8 @@ from dagster_utils.resources.data_repo.jade_data_repo import jade_data_repo_clie
 from dagster_utils.resources.google_storage import google_storage_client
 from dagster_utils.resources.slack import live_slack_client
 
+from hca_orchestration.config.dev_refresh.dev_refresh import run_config_for_per_project_dataset_partition, \
+    run_config_for_cut_snapshot_partition
 from hca_orchestration.contrib.dagster import configure_partitions_for_pipeline
 from hca_orchestration.config import preconfigure_resource_for_mode
 from hca_orchestration.config.dcp_release.dcp_release import run_config_for_dcp_release_partition
@@ -23,6 +25,7 @@ from hca_orchestration.resources.config.dagit import dagit_config
 from hca_orchestration.resources.config.scratch import scratch_config
 from hca_orchestration.resources.config.target_hca_dataset import target_hca_dataset
 from hca_orchestration.resources.data_repo_service import data_repo_service
+from hca_orchestration.repositories.common import copy_project_to_new_dataset_job
 
 
 def validate_ingress_job() -> PipelineDefinition:
@@ -57,12 +60,16 @@ def load_hca_job() -> PipelineDefinition:
 @repository
 def all_jobs() -> list[PipelineDefinition]:
     jobs = [
-        load_hca_job(),
-        validate_ingress_job(),
-        legacy_cut_snapshot_job("prod", "monster@firecloud.org"),
+        copy_project_to_new_dataset_job(),
         cut_project_snapshot_job("prod", "prod", "monster@firecloud.org"),
-        cut_project_snapshot_job("prod", "real_prod", "monster@firecloud.org")
+        cut_project_snapshot_job("prod", "real_prod", "monster@firecloud.org"),
+        legacy_cut_snapshot_job("prod", "monster@firecloud.org"),
+        load_hca_job(),
+        validate_ingress_job()
     ]
+    jobs += configure_partitions_for_pipeline("copy_project_to_new_dataset",
+                                              run_config_for_per_project_dataset_partition)
+    jobs += configure_partitions_for_pipeline("cut_snapshot", run_config_for_cut_snapshot_partition)
     jobs += configure_partitions_for_pipeline("load_hca", run_config_for_dcp_release_partition)
     jobs += configure_partitions_for_pipeline("validate_ingress", run_config_for_validation_ingress_partition)
