--- orchestration/workspace.yaml
@@ -1,6 +1,0 @@
-# Defines where to find our code for tooling like Dagit
-# https://docs.dagster.io/overview/repositories-workspaces/workspaces
-load_from:
-  - python_package:
-      package_name: hca_orchestration.repositories
-      attribute: all_jobs
